+++ fractal/parser.py
@@ -731,6 +731,11 @@ user_register_parser.add_argument(
     required=False,
 )
 user_register_parser.add_argument(
+    "--username",
+    help="Username associated to the user",
+    required=False,
+)
+user_register_parser.add_argument(
     "--superuser",
     help="Give superuser privileges to the new user",
     action="store_true",
@@ -777,7 +782,9 @@ user_edit_parser.add_argument(
 user_edit_parser.add_argument(
     "--new-slurm-user", help="New SLURM username", type=str, required=False
 )
-
+user_edit_parser.add_argument(
+    "--new-username", help="New user username", type=str, required=False
+)
 user_edit_parser_superuser = user_edit_parser.add_mutually_exclusive_group()
 user_edit_parser_superuser.add_argument(
     "--make-superuser",