+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix a bug in tests, by starting the fractal-server FastAPI app in a more standard way (\#481).
 * Require pydantic version to be `>=1.10.8` (\#486, \#490).
 * Remove `--private` option for `fractal task collect` (\#493).
+* Add `--username/--new-username` to `fractal user` subcommands (\#493).
 
 # 1.2.0